+++ packages/evokit-box/CHANGELOG.md
@@ -1,3 +1,5 @@
+[create_theme]: /docs/base/theme
+
 # EvoKit - Box
 
 [![](https://img.shields.io/npm/v/evokit-box.svg)](https://www.npmjs.com/package/evokit-box)
@@ -5,6 +7,17 @@
 ---
 
 ### CHANGELOG
+
+#### 3.0.2 *(11.07.2019)*
+
+- **Added** new advanced border props:
+    - `box-border-top`
+    - `box-border-right`
+    - `box-border-bottom`
+    - `box-border-left`
+    - `box-border-tb`
+    - `box-border-lr`
+- Please [update theme][create_theme]
 
 #### 3.0.1 *(20.06.2019)*